feat: add function on update for board

+++ src/main/java/com/nowgnodeel/boardbe/board/service/BoardService.java
@@ -1,12 +1,14 @@
 package com.nowgnodeel.boardbe.board.service;
 
 import com.nowgnodeel.boardbe.board.dto.CreateBoardRequestDto;
+import com.nowgnodeel.boardbe.board.dto.UpdateBoardRequestDto;
 import com.nowgnodeel.boardbe.board.entity.Board;
 import com.nowgnodeel.boardbe.board.repository.BoardRepository;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.server.ResponseStatusException;
 
 @Service
 @RequiredArgsConstructor
@@ -15,22 +17,28 @@ public class BoardService {
     private final BoardRepository boardRepository;
 
     @Transactional
-    public ResponseEntity<String> createBoard(CreateBoardRequestDto requestDto) {
+    public Board createBoard(CreateBoardRequestDto requestDto) {
         Board board = Board.builder()
                 .title(requestDto.title())
                 .content(requestDto.content())
                 .category(requestDto.category())
                 .build();
-
-        boardRepository.save(board);
-
-        return ResponseEntity.ok().body("Board created successfully");
+        return boardRepository.save(board);
     }
 
     @Transactional
-    public ResponseEntity<String> deleteBoard(Long id) {
+    public void deleteBoard(Long id) {
+        if (!boardRepository.existsById(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Board not found");
+        }
         boardRepository.deleteById(id);
+    }
 
-        return ResponseEntity.ok().body("Board deleted successfully");
+    @Transactional
+    public Board updateBoard(Long id, UpdateBoardRequestDto requestDto) {
+        Board board = boardRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Board not found"));
+        board.patch(requestDto);
+        return board;
     }
 }
